pyproject.toml
[tool.poetry]
name = "nonebot-adapter-console"
version = "0.1.0-beta.1"
description = "Console Adapter for NoneBot2"
authors = ["SuCicada <pengyifu@gmail.com>"]
license = "MIT"
readme = "README.md"
#homepage = "https://v2.nonebot.dev/"
repository = "https://github.com/SuCicada/nonebot-adapter-console"
#documentation = "https://v2.nonebot.dev/"
keywords = ["bot", "console"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
packages = [{ include = "nonebot" }]

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0b1"

[tool.poetry.dev-dependencies]
websocket-client = "^1.3.2"
rel = "^0.4.7"

#[tool.poetry.group.dev.dependencies]
#nonebot2 = { git = "https://github.com/nonebot/nonebot2.git" }
#
#[tool.poetry.group.formatter.dependencies]
#isort = "^5.10.1"
#black = "^22.3.0"
#pre-commit = "^2.17.0"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 80
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
